ECM-73 - Refactor
