+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/helpers/Helper.java
@@ -11,17 +11,14 @@ import uk.gov.hmcts.ecm.common.model.ccd.items.DateListedTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.HearingTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.JurCodesTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.RespondentSumTypeItem;
-import uk.gov.hmcts.ecm.common.model.ccd.items.RepresentedTypeRItem;
 import uk.gov.hmcts.ecm.common.model.ccd.types.CorrespondenceScotType;
 import uk.gov.hmcts.ecm.common.model.ccd.types.CorrespondenceType;
 import uk.gov.hmcts.ecm.common.model.ccd.types.DateListedType;
-import uk.gov.hmcts.ecm.common.model.ccd.types.RepresentedTypeR;
 import uk.gov.hmcts.ecm.common.model.labels.LabelPayloadES;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.ListIterator;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -349,52 +346,4 @@ public class Helper {
                 : new ArrayList<>();
     }
 
-    public static CaseData populateDynamicRespondentRepresentativeNames(CaseData caseData) {
-        List<DynamicValueType> listItems = createDynamicRespondentName(caseData.getRespondentCollection());
-        if (!listItems.isEmpty()) {
-            var dynamicFixedListType = new DynamicFixedListType();
-            dynamicFixedListType.setListItems(listItems);
-            if (caseData.getRepCollection() != null && !caseData.getRepCollection().isEmpty()) {
-                ListIterator<RepresentedTypeRItem> repItr = caseData.getRepCollection().listIterator();
-                int index;
-                while (repItr.hasNext()) {
-                    index = repItr.nextIndex() + 1;
-                    var respRepCollection = caseData.getRepCollection().get(index - 1);
-                    var dynamicValueType = new DynamicValueType();
-                    if (respRepCollection.getValue().getDynamicRespRepName() == null) {
-                        repItr.next().getValue().setDynamicRespRepName(dynamicFixedListType);
-                        var respRepName = respRepCollection.getValue().getRespRepName();
-                        dynamicValueType.setLabel(respRepName);
-                        dynamicValueType.setCode(respRepName);
-                    } else {
-                        dynamicValueType = respRepCollection.getValue().getDynamicRespRepName().getValue();
-                        repItr.next().getValue().setDynamicRespRepName(dynamicFixedListType);
-                    }
-                    respRepCollection.getValue().getDynamicRespRepName().setValue(dynamicValueType);
-                }
-            } else{
-                var representedTypeR = new RepresentedTypeR();
-                representedTypeR.setDynamicRespRepName(dynamicFixedListType);
-                var representedTypeRItem = new RepresentedTypeRItem();
-                representedTypeRItem.setValue(representedTypeR);
-                var collection = List.of(representedTypeRItem);
-                caseData.setRepCollection(collection);
-            }
-        }
-        return caseData;
-    }
-
-    private static List<DynamicValueType> createDynamicRespondentName(List<RespondentSumTypeItem> respondentCollection) {
-        List<DynamicValueType> listItems = new ArrayList<>();
-        if (respondentCollection != null) {
-            for (RespondentSumTypeItem respondentSumTypeItem : respondentCollection) {
-                var dynamicValueType = new DynamicValueType();
-                var respondentSumType = respondentSumTypeItem.getValue();
-                dynamicValueType.setCode(respondentSumType.getRespondentName());
-                dynamicValueType.setLabel(respondentSumType.getRespondentName());
-                listItems.add(dynamicValueType);
-            }
-        }
-        return listItems;
-    }
 }
